builtin_cd listo. Se añade otro modulo auxiliar para modificar variables entorno. Tambien se renombra el builtin_env a init_environment ya que este modulo solo guarda en memoria dinamica el valor de las variables de entorno al iniciar la ejecucion
