+++ Makefile
@@ -6,7 +6,7 @@
 #    By: alvicina <alvicina@student.42urduliz.co    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/01/06 17:48:55 by alejandro         #+#    #+#              #
-#    Updated: 2024/01/09 19:10:57 by alvicina         ###   ########.fr        #
+#    Updated: 2024/01/11 11:18:06 by alvicina         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -25,7 +25,7 @@ HEADER_SRCS = minishell.h
 HEADER_DIR  = includes/
 HEADER      = $(addprefix $(HEADER_DIR), $(HEADER_SRCS))
 
-MPATH_SRCS  = main.c sig_handler.c cmd_preprocessor.c ast_builder.c ast_processor.c builtin_env.c builtin_pwd.c builtin_cd.c
+MPATH_SRCS  = main.c sig_handler.c cmd_preprocessor.c ast_builder.c ast_processor.c init_environment.c builtin_pwd.c builtin_cd.c environment_utils.c
 
 MPATH_DIR   = mandatory/
 MPATH       = $(addprefix $(MPATH_DIR), $(MPATH_SRCS))
