+++ mandatory/builtin_cd.c
@@ -6,81 +6,78 @@
 /*   By: alvicina <alvicina@student.42urduliz.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/09 13:12:45 by alvicina          #+#    #+#             */
-/*   Updated: 2024/01/10 20:05:44 by alvicina         ###   ########.fr       */
+/*   Updated: 2024/01/11 11:16:34 by alvicina         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
 
-int ft_exec_update_env(t_mshell *mini_data, char *which, char *to_change)
+static int	do_cd_exec(char *cwd_be, t_mshell *mini_data,
+char *arguments, int free_b)
+{
+	char	*cwd_af;
+
+	if (chdir(arguments))
+		return (perror("cd: could not change to directory"), 4);
+	if (ft_update_env(cwd_be, mini_data, "OLDPWD") == -1)
+		return (ft_putstr_fd("cd: could not update env", 2), 3);
+	cwd_af = getcwd(NULL, 0);
+	if (cwd_af == NULL)
+		return (ft_putstr_fd("cd: could not update env", 2), 3);
+	if (ft_update_env(cwd_af, mini_data, "PWD") == -1)
+		return (ft_putstr_fd("cd: could not update env", 2), 3);
+	if (free_b)
+		free(arguments);
+	free(cwd_af);
+	return (0);
+}
+
+static char	*set_cd_one_up(char *cwd)
 {
 	size_t	i;
-	char	*temp;
+	char	*dir_one_up;
 
 	i = 0;
-	while (mini_data->env_custom[i])
+	while (cwd[i])
+		i++;
+	i = i - 1;
+	while (cwd[i] != '/' && cwd[i])
+		i--;
+	dir_one_up = ft_substr(cwd, 0, ft_strlen(cwd) - ft_strlen(&cwd[i]));
+	if (dir_one_up == NULL)
+		return (NULL);
+	return (dir_one_up);
+}
+
+static char	*set_cd_special_case(char *cwd, char *arguments,
+t_mshell *mini_data)
+{
+	char	*dir_to_go;
+
+	if (!ft_strncmp(arguments, "..", ft_strlen(arguments)))
 	{
-		if (!ft_strncmp(which, mini_data->env_custom[i], ft_strlen(which) - 1))
-		{
-			temp = mini_data->env_custom[i];
-			mini_data->env_custom[i] = to_change;
-			free(temp);
-			return (0);
-		}
-		i++;
+		dir_to_go = set_cd_one_up(cwd);
+		if (dir_to_go == NULL)
+			return (NULL);
 	}
-	return (-1);
-}
-
-int	ft_update_env(char	*new_content, t_mshell *mini_data, char *which)
-{
-	char	*set_new;
-	char	*to_change;
-	char	*temp;
-	
-	set_new = ft_strdup(new_content);
-	if (set_new == NULL)
-		return (-1);
-	temp = ft_strjoin(which, "=");
-	if (temp == NULL)
-		return (-1);
-	to_change = ft_strjoin(temp, set_new);
-	//printf("%s\n", to_change);
-	if (to_change == NULL)
-		return (-1);
-	free(temp);
-	if (ft_exec_update_env(mini_data, which, to_change) == -1)
-		return (-1);
-	return (free(set_new), 0);
-}
-
-char	*ft_set_env(char *env_to_set)
-{
-	size_t	i;
-
-	i = 0;
-	while (env_to_set[i] && env_to_set[i] != '=')
-		i++;
-	if (env_to_set[i] == '=')
-		i++;
-	return (&env_to_set[i]);
-}
-
-char	*ft_get_env(char *env_to_get, char **envp)
-{
-	size_t	i;
-
-	i = 0;
-	while (envp[i])
+	if (!ft_strncmp(arguments, "-", ft_strlen(arguments)))
 	{
-		if (ft_strncmp(env_to_get, envp[i], ft_strlen(env_to_get) - 1) == 0)
-			break ;
-		i++;
+		dir_to_go = ft_get_env("OLDPWD", mini_data->env_custom);
+		if (dir_to_go == NULL)
+			return (ft_putstr_fd("cd: could not get OLDPWD\n", 2), NULL);
+		dir_to_go = ft_set_env(dir_to_go);
+		if (dir_to_go == NULL)
+			return (ft_putstr_fd("cd: could not set OLDPWD\n", 2), NULL);
+		dir_to_go = ft_strdup(dir_to_go);
+		if (dir_to_go == NULL)
+			return (ft_putstr_fd("cd: error getting OLPWD value", 2), NULL);
+		ft_putstr_fd(dir_to_go, 1);
+		write(1, "\n", 1);
 	}
-	return (envp[i]);
+	return (dir_to_go);
 }
 
-static int do_cd_home(t_mshell *mini_data)
+static int	do_cd_home(t_mshell *mini_data)
 {
 	char	*dir_to_save;
 	char	*env_home;
@@ -93,7 +90,7 @@ static int do_cd_home(t_mshell *mini_data)
 		return (ft_putstr_fd("cd: could not set home", 2), 2);
 	if (ft_update_env(env_home, mini_data, "PWD") == -1)
 		return (ft_putstr_fd("cd: could not update env", 2), 3);
-	dir_to_save = getcwd(NULL, 0); //tengo que liberar
+	dir_to_save = getcwd(NULL, 0);
 	if (dir_to_save == NULL)
 		return (ft_putstr_fd("cd: could not update env", 2), 3);
 	if (ft_update_env(dir_to_save, mini_data, "OLDPWD") == -1)
@@ -114,14 +111,14 @@ int	do_cd(t_mshell *mini_data, char **arguments)
 	free_b = 0;
 	if (!ft_strncmp(arguments[i], ".", ft_strlen(arguments[i])))
 		return (0);
-	if (!ft_strncmp(arguments[i], "--", 3) 
-	|| !ft_strncmp(arguments[i], "~", ft_strlen(arguments[i])))
+	if (!ft_strncmp(arguments[i], "--", 3)
+		|| !ft_strncmp(arguments[i], "~", ft_strlen(arguments[i])))
 		return (do_cd_home(mini_data));
 	if (arguments[i] == NULL)
 		return (do_cd_home(mini_data));
 	cwd = getcwd(NULL, 0);
-	if (!ft_strncmp(arguments[i], "..", ft_strlen(arguments[i])) 
-	|| !ft_strncmp(arguments[i], "-", ft_strlen(arguments[i])))
+	if (!ft_strncmp(arguments[i], "..", ft_strlen(arguments[i]))
+		|| !ft_strncmp(arguments[i], "-", ft_strlen(arguments[i])))
 	{
 		arguments[i] = set_cd_special_case(cwd, arguments[i], mini_data);
 		if (arguments[i] == NULL)
@@ -131,7 +128,7 @@ int	do_cd(t_mshell *mini_data, char **arguments)
 	free(cwd);
 	return (do_cd_exec(cwd, mini_data, arguments[i], free_b));
 }
-
+/*
 int	main(int argc, char **argv, char **envp)
 {
 	t_mshell	mini_data;
@@ -143,26 +140,28 @@ int	main(int argc, char **argv, char **envp)
 	(void) argc;
 	(void) argv;
 	arguments[0] = "cd";
-	arguments[1] = "../libft";
+	arguments[1] = "..";
 	arguments[2] = NULL;
-	mini_data.env_custom = do_env(envp, 0);
-	/*while (mini_data.env_custom[i])
+	mini_data.env_custom = do_env_init(envp, 0);
+	while (mini_data.env_custom[i])
 	{
 		printf("%s\n", mini_data.env_custom[i]);
 		i++;
-	}*/
+	}
 	i = 0;
 	printf("********************************************************\n");
 	do_cd(&mini_data, arguments);
 	printf("********************************************************\n");
-	/*while (mini_data.env_custom[i])
+	while (mini_data.env_custom[i])
 	{
 		printf("%s\n", mini_data.env_custom[i]);
 		i++;
-	}*/
+	}
 	cwd = getcwd(NULL, 0);
 	printf("********************************************************\n");
 	printf("cwd: %s\n", cwd);
+	printf("********************************************************\n");
+	free(cwd);
 	ft_free_env(mini_data.env_custom);
 	return (0);
-}
+}*/
