+++ mandatory/builtin_pwd.c
@@ -6,7 +6,7 @@
 /*   By: alvicina <alvicina@student.42urduliz.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/08 15:37:55 by alvicina          #+#    #+#             */
-/*   Updated: 2024/01/08 16:32:29 by alvicina         ###   ########.fr       */
+/*   Updated: 2024/01/11 10:40:39 by alvicina         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 // NOTA: getcwd asigna memoria dinamica. Hay que hacer free al retorno una vez
 // no haga fala.
 // NOTA: de momento este modulo imprime y libera el resultado. Tal vez podriamos
-// necesitar que devolviera el resultado para este ser gestionado (imprimido y liberado)
-// desde otro lugar;
+// necesitar que devolviera el resultado para este ser gestionado 
+//(imprimido y liberado) desde otro lugar;
 
 static void	print_pwd(char *cw_directory)
 {
